requirements.txt
fastapi
uvicorn
httpx
python-multipart
pydantic
PyJWT[crypto]
casparser
openpyxl
xlrd>=2.0.1
